Refactored utils into separate file
