+++ package.json
@@ -27,6 +27,7 @@
     "crypto-js": "^3.1.9-1",
     "cryptojs-extension": "pkaminski/cryptojs-extension",
     "firebase": "^2.4.2",
+    "lru-cache": "^4.1.3",
     "serialized-lru-cache": "^3.1.0"
   }
 }
